# Track balances as Number64UintType