--- a/packages/beacon-state-transition/src/altair/block/processAttestation.ts
+++ b/packages/beacon-state-transition/src/altair/block/processAttestation.ts
@@ -100,7 +100,7 @@ export function processAttestations(
     proposerReward += proposerRewardNumerator / PROPOSER_REWARD_DOMINATOR;
   }
 
-  increaseBalance(state, epochCtx.getBeaconProposer(state.slot), proposerReward);
+  increaseBalance(state, epochCtx.getBeaconProposer(state.slot), Number(proposerReward));
 }
 
 /**