… packages/api/package.json
… packages/api/test/unit/beacon.test.ts
… packages/beacon-state-transition/package.json
… packages/beacon-state-transition/src/altair/epoch/balance.ts
… packages/beacon-state-transition/src/util/balance.ts
… packages/beacon-state-transition/src/util/genesis.ts
… packages/beacon-state-transition/test/perf/allForks/hashing.test.ts
… packages/beacon-state-transition/test/perf/phase0/epoch/processEffectiveBalanceUpdates.test.ts
… packages/beacon-state-transition/test/perf/util.ts
… packages/beacon-state-transition/test/unit/util/balance.test.ts
… packages/beacon-state-transition/test/utils/state.ts
… packages/cli/package.json
… packages/config/package.json
… packages/db/package.json
… packages/fork-choice/package.json
… packages/light-client/package.json
… packages/light-client/test/lightclientMockServer.ts
… packages/light-client/test/utils.ts
… packages/lodestar/package.json
… packages/lodestar/src/api/impl/beacon/state/utils.ts
… packages/lodestar/test/unit/api/impl/beacon/state/stateValidators.test.ts
… packages/lodestar/test/unit/chain/forkChoice/forkChoice.test.ts
… packages/lodestar/test/utils/balances.ts
… packages/lodestar/test/utils/state.ts
… packages/params/package.json
… packages/spec-test-runner/package.json
… packages/spec-test-runner/test/spec/altair/rewards.test.ts
… packages/spec-test-util/package.json
… packages/types/package.json
… packages/types/src/altair/sszTypes.ts