**Motivation**

Track `balances` as the new `Number64UintType` of ssz for faster processing, same for slot, epoch

**Description**

This depends on https://github.com/ChainSafe/ssz/pull/159

+ For all Number64 type, use the new `Number64UintType` of ssz
+ Track balances as a number too `new ListType({elementType: Number64, limit: VALIDATOR_REGISTRY_LIMIT})`
+ Update CachedBalancesList: remove `persistent-ts` array there. **TODO**: remove proxy also, may address that in a separate issue.
+ Increase balance/decrease balance: use the new api `tree_applyUint64Delta` to avoid the pattern `balances[i] = balances[i] + delta` (which have to do `deserialize` then `serialize`)
+ `processRewardsAndPenalties`: use the new api `tree_newTreeFromUint64Deltas` to build balances tree and avoid BigInt calculation and Uint8Array allocation 

Closes #2972 

Along with the new `persistent-merkle-tree` and `as-sha256` improvements, we're likely to double the speed of epoch transition  as I tested in my local environment

|#|master|new branch|
|-|-------|------------|
|beforeProcessEpoch|1.603897  s/op|697.9248 ms/op|
|processRewardsAndPenalties|704.3871 ms/op|138.6842 ms/op|

More related improvement could be done after this PR:
+ Track effectiveBalance as number too, may simplify calculation at different places a bit
+ Remove CachedBalanceList proxy