@tuyennhv 2021-08-25
> Killer improvements!! 🔥 amazing job

just want to clarify that most of the improvement come from the new ssz and we already merge that to master. The main improvement for this PR is `processRewardsAndPenalties` and removing a separate `balances` persistent-ts cache