@tuyennhv 2021-08-31
| #  |this branch | master | ratio|
| -- | -- | -- | -- |
| pyrmont_e62330 - altair processRewardsAndPenalties | 78.252 ms/op | 213.63 ms/op | 0.37 |
| altair processRewardsAndPenalties - 250000 normalcase | 123.93 ms/op | 338.31 ms/op | 0.37 |
| altair processRewardsAndPenalties - 250000 worstcase | 139.13 ms/op | 281.75 ms/op | 0.49 |
| mainnet_e58758 - phase0 processRewardsAndPenalties | 102.81 ms/op | 273.29 ms/op | 0.38 |